Refresh statsd client periodically
Right now, the StatsdClient is only instantiated once (the first time
`statsd` is imported). This poses a problem when the statsd host changes
IP address, because the DNS won't be looked up again (and since we use
UDP, we won't realise our calls aren't arriving either).

So, we now periodically refresh the client (every two minutes by
default). This is a compromise of doing it on every call (which would
dramatically increase the load of the logging), which means we will only
lose a little bit of stats.

This approach also requires changing the mocking in the tests a bit.